requirements.txt
Django==1.9
Pillow==3.1.0
argparse==1.2.1
certifi==2017.4.17
chardet==3.0.4
dj-database-url==0.4.2
django-allauth==0.32.0
django-crispy-forms==1.6.0
django-disqus==0.5
django-markdown-deux==1.0.5
django-material==1.0.0
django-pagedown==0.1.1
funcsigs==1.0.2
gunicorn==19.7.1
idna==2.5
markdown2==2.3.1
mock==2.0.0
oauthlib==2.0.2
pbr==3.1.1
psycopg2==2.7.1
python-openid==2.2.5
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
urllib3==1.21.1
whitenoise==3.3.0
wsgiref==0.1.2
